patch: adds frameworks/base
This patch adds patch for frameworks/base.

When AOSP android booting, launcher runs twice.
One is in Setting(Fake) and other is real.
Real launcher can run after user booting is done and user booting is
dependent on several parts.

The aim of this patch is to run real launcher at once.

Change-Id: I97f57a5fb480c31adc3053eec8e5c01bb60bfdd5
Signed-off-by: Sungwoo Park <swpark@nexell.co.kr>